gh-132: Extract Root
